+++ Vitrine/Settings/SettingsCodec.swift
@@ -46,6 +46,7 @@ enum SettingsCodec {
         static let exportFormat = "exportFormat"
         static let colorProfile = "colorProfile"
         static let richClipboard = "richClipboard"
+        static let concealClipboard = "concealClipboard"
         static let textSidecar = "textSidecar"
         static let hotkeyAction = "hotkeyAction"
         static let appLanguage = "appLanguage"
@@ -109,7 +110,8 @@ enum SettingsCodec {
             showShadow, showLineNumbers, wrapColumns, highlightedLines, focusHighlightedLines,
             diffDecorations, annotations, metadata, gradientPreset,
             backgroundStyle, autoCopy, alsoSaveToFile, closeAfterCopy, exportScale, exportFormat,
-            colorProfile, richClipboard, textSidecar, hotkeyAction, appLanguage, treatURLs,
+            colorProfile, richClipboard, textSidecar, concealClipboard, hotkeyAction, appLanguage,
+            treatURLs,
             reindentOnPaste,
             webViewportKind, webViewports, webCustomViewportWidth, webCustomViewportHeight,
             webCaptureMode, webWaitKind, webWaitSeconds, webUsesLoggedInSession,